move dummy_encoder_helper to utils
