+++ server/website/tests/test_tasks.py
@@ -1,3 +1,8 @@
+#
+# OtterTune - test_tasks.py
+#
+# Copyright (c) 2017-18, Carnegie Mellon University Database Group
+#
 import copy
 import numpy as np
 
@@ -9,11 +14,12 @@ from website.models import Workload, PipelineRun, PipelineData
 from website.tasks.periodic_tasks import (run_background_tasks,
                                           aggregate_data,
                                           run_workload_characterization,
-                                          run_knob_identification,
-                                          dummy_encoder_helper)
+                                          run_knob_identification)
 from website.types import PipelineTaskType
 
-@override_settings(CELERY_ALWAYS_EAGER = True, TEST_RUNNER = 'djcelery.contrib.test_runner.CeleryTestSuiteRunner')
+
+@override_settings(CELERY_ALWAYS_EAGER=True,
+                   TEST_RUNNER='djcelery.contrib.test_runner.CeleryTestSuiteRunner')
 class BackgroundTestCase(TestCase):
 
     fixtures = ['test_website.json']
@@ -56,6 +62,7 @@ class BackgroundTestCase(TestCase):
     def testNewRankedKnobs(self):
         self.checkNewTask(PipelineTaskType.RANKED_KNOBS)
 
+
 class AggregateTestCase(TestCase):
 
     fixtures = ['test_website.json']
@@ -69,6 +76,7 @@ class AggregateTestCase(TestCase):
             for k in keys:
                 self.assertIn(k, d)
 
+
 class PrunedMetricTestCase(TestCase):
 
     fixtures = ['test_website.json']
@@ -80,6 +88,7 @@ class PrunedMetricTestCase(TestCase):
         for m in pruned_metrics:
             self.assertIn(m, metric_data['columnlabels'])
 
+
 class RankedKnobTestCase(TestCase):
 
     fixtures = ['test_website.json']
@@ -88,7 +97,8 @@ class RankedKnobTestCase(TestCase):
         workloads = Workload.objects.all()
         knob_data, metric_data = aggregate_data(workloads[0])
 
-        # instead of doing actual metric pruning by factor analysis / clustering, just randomly select 5 nonconstant metrics
+        # instead of doing actual metric pruning by factor analysis / clustering,
+        # just randomly select 5 nonconstant metrics
         nonconst_metric_columnlabels = []
         for col, cl in zip(metric_data['data'].T, metric_data['columnlabels']):
             if np.any(col != col[0]):
@@ -96,9 +106,8 @@ class RankedKnobTestCase(TestCase):
 
         num_metrics = min(5, len(nonconst_metric_columnlabels))
         selected_columnlabels = choice(nonconst_metric_columnlabels, num_metrics, replace=False)
-        pruned_metric_idxs = [i for i, metric_name in
-            enumerate(metric_data['columnlabels'])
-            if metric_name in selected_columnlabels]
+        pruned_metric_idxs = [i for i, metric_name in enumerate(metric_data['columnlabels'])
+                              if metric_name in selected_columnlabels]
         pruned_metric_data = {
             'data': metric_data['data'][:, pruned_metric_idxs],
             'rowlabels': copy.deepcopy(metric_data['rowlabels']),
@@ -109,33 +118,3 @@ class RankedKnobTestCase(TestCase):
         ranked_knobs = run_knob_identification(knob_data, pruned_metric_data)
         for k in ranked_knobs:
             self.assertIn(k, knob_data['columnlabels'])
-
-
-class DummyEncoderHelperTestCase(TestCase):
-
-    fixtures = ['postgres-96_knobs.json']
-
-    def test_no_featured_categorical(self):
-        featured_knobs = ['global.backend_flush_after',
-                          'global.bgwriter_delay',
-                          'global.wal_writer_delay',
-                          'global.work_mem']
-        categorical_info = dummy_encoder_helper(featured_knobs)
-        self.assertEqual(len(categorical_info['n_values']), 0)
-        self.assertEqual(len(categorical_info['categorical_features']), 0)
-        self.assertEqual(categorical_info['cat_columnlabels'], [])
-        self.assertEqual(categorical_info['noncat_columnlabels'], featured_knobs)
-
-    def test_featured_categorical(self):
-        featured_knobs = ['global.backend_flush_after',
-                          'global.bgwriter_delay',
-                          'global.wal_writer_delay',
-                          'global.work_mem',
-                          'global.wal_sync_method']  # last knob categorical
-        categorical_info = dummy_encoder_helper(featured_knobs)
-        self.assertEqual(len(categorical_info['n_values']), 1)
-        self.assertEqual(categorical_info['n_values'][0], 4)
-        self.assertEqual(len(categorical_info['categorical_features']), 1)
-        self.assertEqual(categorical_info['categorical_features'][0], 4)
-        self.assertEqual(categorical_info['cat_columnlabels'], ['global.wal_sync_method'])
-        self.assertEqual(categorical_info['noncat_columnlabels'], featured_knobs[:-1])
